Create ECR repos for the CI images
This adds some repositories to enable https://github.com/apache/tvm/issues/10646. The main caveat with this approach (vs letting CI nodes create the repositories as necessary) is that new images require an interaction with the infra team, but this is probably fine since these are pretty rare (`ci_hexagon` is the only new one in the past several months).