Add S3 bucket for CI artifacts
This adds a bucket which would be used by CI builds to store
intermediate artifacts rather than Jenkins' built in `stash` and
`unstash` directives. Those take a lot of time on some occasions (20~ish
minutes) and slow down the Jenkins head node (since it has to do the
actual storing of data).

The counterpart in apache/tvm would replace `stash` with a function to
store data in S3 instead.